fix: use bigint in database

+++ apps/url-shortener/prisma/migrations/20250616111917_init/migration.sql
@@ -1,6 +1,6 @@
 -- CreateTable
 CREATE TABLE "shortend_urls" (
-    "id" INTEGER NOT NULL,
+    "id" BIGINT NOT NULL,
     "url" TEXT NOT NULL,
 
     CONSTRAINT "shortend_urls_pkey" PRIMARY KEY ("id")
